Fetch dataset version back
