+++ client/verta/verta/dataset/_dataset_blob.py
@@ -283,123 +283,3 @@ class _DatasetBlob(blob.Blob):
         """
         components = self._components_map.values()
         return list(sorted(components, key=lambda component: component.path))
-
-
-class Component(object):
-    """
-    A dataset component returned by ``dataset.list_components()``.
-
-    Attributes
-    ----------
-    path : str
-        File path.
-    base_path : str
-        Prefix of `path`.
-    size : int
-        File size.
-    last_modified : int
-        Unix time when this file was last modified.
-    sha256 : str
-        SHA-256 checksum.
-    md5 : str
-        MD5 checksum.
-
-    """
-    def __init__(
-            self,
-            path, size=None, last_modified=None,
-            sha256=None, md5=None,
-            base_path=None,
-            internal_versioned_path=None, local_path=None):
-        # metadata
-        self.path = path
-        self.size = size
-        self.last_modified = last_modified
-
-        # checksums
-        self.sha256 = sha256
-        self.md5 = md5
-
-        # base path
-        self.base_path = base_path
-
-        # ModelDB versioning
-        self._internal_versioned_path = internal_versioned_path
-        self._local_path = local_path
-
-    def __repr__(self):
-        lines = [self.path]
-
-        if self.base_path:
-            lines.append("base path: {}".format(self.base_path))
-        if self.size:
-            lines.append("{} bytes".format(self.size))
-        if self.last_modified:
-            lines.append("last modified: {}".format(_utils.timestamp_to_str(self.last_modified)))
-        if self.sha256:
-            lines.append("SHA-256 checksum: {}".format(self.sha256))
-        if self.md5:
-            lines.append("MD5 checksum: {}".format(self.md5))
-
-        return "\n    ".join(lines)
-
-    def __eq__(self, other):
-        if not isinstance(other, type(self)):
-            return NotImplemented
-
-        return self.__dict__ == other.__dict__
-
-    @classmethod
-    def _from_proto(cls, component_msg):
-        return cls(
-            path=component_msg.path,
-            size=component_msg.size or None,
-            last_modified=component_msg.last_modified_at_source or None,
-            sha256=component_msg.sha256 or None,
-            md5=component_msg.md5 or None,
-            internal_versioned_path=component_msg.internal_versioned_path or None,
-            base_path=component_msg.base_path,
-        )
-
-    def _as_proto(self):
-        return _DatasetService.PathDatasetComponentBlob(
-            path=self.path or "",
-            size=self.size or 0,
-            last_modified_at_source=self.last_modified or 0,
-            sha256=self.sha256 or "",
-            md5=self.md5 or "",
-            internal_versioned_path=self._internal_versioned_path or "",
-            base_path=self.base_path or "",
-        )
-
-
-class S3Component(Component):
-    def __init__(self, s3_version_id=None, **kwargs):
-        super(S3Component, self).__init__(**kwargs)
-
-        # S3 versioning
-        self.s3_version_id = s3_version_id
-
-    def __repr__(self):
-        repr_str = super(S3Component, self).__repr__()
-        lines = repr_str.split("\n    ")
-
-        if self.s3_version_id:
-            lines.append("S3 version ID: {}".format(self.s3_version_id))
-
-        return "\n    ".join(lines)
-
-
-    @classmethod
-    def _from_proto(cls, s3_component_msg):
-        obj = super(S3Component, cls)._from_proto(s3_component_msg.path)
-        obj.s3_version_id = s3_component_msg.s3_version_id
-
-        return obj
-
-    def _as_proto(self):
-        s3_component_msg = _DatasetService.S3DatasetComponentBlob()
-        s3_component_msg.path.CopyFrom(super(S3Component, self)._as_proto())
-        s3_component_msg.s3_version_id = self.s3_version_id or ""
-
-        return s3_component_msg
